# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Go 语言构建的轻量级、高性能通用电商后端系统。",
        "title": "GoShop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3233",
    "basePath": "/",
    "paths": {
...
        "/api/cart": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "获取购物车",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "添加或更新购物车",
                "parameters": [
                    {
                        "description": "购物车项",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CartSaveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cart/sync": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "同步购物车",
                "parameters": [
                    {
                        "description": "同步项",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CartSyncReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cart/{skuId}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "删除购物车项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SKU ID",
                        "name": "skuId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/user-coupons/receive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotion"
                ],
                "summary": "领取优惠券",
                "parameters": [
                    {
                        "description": "领取参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ReceiveCouponReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CartSaveReq": {
            "type": "object",
            "required": [
                "quantity",
                "skuId"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "skuId": {
                    "type": "integer"
                }
            }
        },
        "handlers.CartSyncReq": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SyncItem"
                    }
                }
            }
        },
...
        "handlers.ReceiveCouponReq": {
            "type": "object",
            "required": [
                "couponId"
            ],
            "properties": {
                "couponId": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SyncItem": {
            "type": "object",
            "required": [
                "quantity",
                "skuId"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "skuId": {
                    "type": "integer"
                }
            }
        },
...
    }
}